+++ lib/ast-transform.js
@@ -83,7 +83,7 @@ function buildAttr(b, name, content) {
   }
 }
 
-module.exports = class AstTransform {
+class AstTransform {
   constructor() {
     this.syntax = null;
   }
@@ -92,30 +92,25 @@ module.exports = class AstTransform {
     let b = this.syntax.builders;
     let traverse = this.syntax.traverse;
 
-    try {
-
-      traverse(ast, {
-        BlockStatement: (node) => {
-          if (node.path.original === 'fa-list') {
-            return this.transformFaListBlock(b, traverse, node);
-          } else if (node.path.original === 'fa-stack') {
-            return this.transformFaStackBlock(b, traverse, node);
-          }
-        },
-
-        MustacheStatement: (node) => {
-          if (node.path.original === 'fa-icon') {
-            return this.transformFaIcon(b, node);
-          } else if (node.path.original === 'fa-list') {
-            return this.buildFaList(b, node);
-          } else if (node.path.original === 'fa-stack') {
-            return this.buildFaStack(b, node);
-          }
+    traverse(ast, {
+      BlockStatement: (node) => {
+        if (node.path.original === 'fa-list') {
+          return this.transformFaListBlock(b, traverse, node);
+        } else if (node.path.original === 'fa-stack') {
+          return this.transformFaStackBlock(b, traverse, node);
         }
-      });
-    } catch(e) {
-      debugger;
-    }
+      },
+
+      MustacheStatement: (node) => {
+        if (node.path.original === 'fa-icon') {
+          return this.transformFaIcon(b, node);
+        } else if (node.path.original === 'fa-list') {
+          return this.buildFaList(b, node);
+        } else if (node.path.original === 'fa-stack') {
+          return this.buildFaStack(b, node);
+        }
+      }
+    });
 
     return ast;
   }
@@ -324,3 +319,14 @@ module.exports = class AstTransform {
     return b.element(tagName, attrs)
   }
 }
+
+function buildAstTransform(addon) {
+  return class EmberFontAwesomeAstTransform extends AstTransform {
+    constructor(options) {
+      super(options);
+      this.addon = addon;
+    }
+  }
+}
+
+module.exports = buildAstTransform;
